Created index.html, images folder, and a CSS folder
Created Index.html, an images folder with one image (add images as we get them), and a CSS folder with a blank stylesheet. 
index.html is our replacement for home.html. I took most of the content from home.html, got rid of any styling, and cleaned up what was left. Copy the nav from this file.
TODO
Update the CV link (it currently takes us to Michaelanne's current site)
Add twitter stream (the original has a lot of javascript/Wordpress that's hard to understand. Let's see if we can do it with HTML and CSS)