Big Footer Update
Added page class to div class = "homepage." Added div class = "skip_t" to let users skip tabbing through the twitter feed.
Added images and contact info to the footer. It's a sticky footer that uses the height of the content above it to stay at the bottom of the page.